+++ docs/conf.py
@@ -4,7 +4,7 @@
 project = "ParamDB"
 copyright = "2023, California Institute of Technology"
 author = "Alex Hadley"
-release = "0.4.0"
+release = "0.5.0"
 
 # General configuration
 extensions = [